fn get_my_address was moved into lp_coins.rs
